feat: create a cloudformation stack for security group deployment
Signed-off-by: David Duncan <davdunc@amazon.com>

The AWS Security team discourages the use of default security
groups. It's an opportunity to be lazy about the configuration. This
transitions over to a least privilege model using two options:
Firstly, by assigning the public ip address for ssh acces (you can't
just open the port to the world) and secondly, by associating a prefix
list with the ssh port that includes the internal network used
throught the Amazon regions. The prefix lists are regionally specific,
so there is a regional mapping in the cloudformation template
template (jinja template that builds the cloudformattion template).